refactor: remove block number configurability
refactor calculation to be with big.Int + update goscale

+++ runtime/balances_force_free_test.go
@@ -53,7 +53,7 @@ func Test_Balances_ForceFree_BadOrigin(t *testing.T) {
 	err = ext.Encode(*encoder)
 	assert.NoError(t, err)
 
-	header := gossamertypes.NewHeader(parentHash, stateRoot, extrinsicsRoot, blockNumber, gossamertypes.NewDigest())
+	header := gossamertypes.NewHeader(parentHash, stateRoot, extrinsicsRoot, uint(blockNumber), gossamertypes.NewDigest())
 	encodedHeader, err := scale.Marshal(*header)
 	assert.NoError(t, err)
