+++ runtime/balances_transfer_test.go
@@ -61,7 +61,7 @@ func Test_Balances_Transfer_Success(t *testing.T) {
 	err = ext.Encode(*encoder)
 	assert.NoError(t, err)
 
-	header := gossamertypes.NewHeader(parentHash, stateRoot, extrinsicsRoot, blockNumber, gossamertypes.NewDigest())
+	header := gossamertypes.NewHeader(parentHash, stateRoot, extrinsicsRoot, uint(blockNumber), gossamertypes.NewDigest())
 	encodedHeader, err := scale.Marshal(*header)
 	assert.NoError(t, err)
 
@@ -168,7 +168,7 @@ func Test_Balances_Transfer_Invalid_InsufficientBalance(t *testing.T) {
 	err = ext.Encode(*encoder)
 	assert.NoError(t, err)
 
-	header := gossamertypes.NewHeader(parentHash, stateRoot, extrinsicsRoot, blockNumber, gossamertypes.NewDigest())
+	header := gossamertypes.NewHeader(parentHash, stateRoot, extrinsicsRoot, uint(blockNumber), gossamertypes.NewDigest())
 	encodedHeader, err := scale.Marshal(*header)
 	assert.NoError(t, err)
 
@@ -229,7 +229,7 @@ func Test_Balances_Transfer_Invalid_ExistentialDeposit(t *testing.T) {
 	err = ext.Encode(*encoder)
 	assert.NoError(t, err)
 
-	header := gossamertypes.NewHeader(parentHash, stateRoot, extrinsicsRoot, blockNumber, gossamertypes.NewDigest())
+	header := gossamertypes.NewHeader(parentHash, stateRoot, extrinsicsRoot, uint(blockNumber), gossamertypes.NewDigest())
 	encodedHeader, err := scale.Marshal(*header)
 	assert.NoError(t, err)
